+++ spec/cmds/docker_spec.js
@@ -5,14 +5,10 @@ h.describeSkipVm("Azk command docker, run", function() {
   var outputs = [];
   var UI   = h.mockUI(beforeEach, outputs);
   var cmd  = init(UI);
-  var argv = process.argv;
-
-  afterEach(() => process.argv = argv);
 
   it("should call `azk vm ssh`", function() {
     cmd.cwd = __dirname;
-    process.argv = [null, null, null, "images"];
-    return cmd.run().then(() => {
+    return cmd.run(["images"]).then(() => {
       h.expect(outputs[0]).to.match(/azk vm ssh -t/);
       h.expect(outputs[0]).to.match(RegExp("cd.*" + h.escapeRegExp(__dirname)));
       h.expect(outputs[0]).to.match(/; docker images/);
@@ -20,8 +16,8 @@ h.describeSkipVm("Azk command docker, run", function() {
   });
 
   it("should forwarding all arguments", function() {
-    process.argv = [null, null, null, "run", "/bin/bash", "-c", "ls -l"];
-    return cmd.run().then(() => {
+    var args = ["run", "/bin/bash", "-c", "ls -l"];
+    return cmd.run(args).then(() => {
       var regex = /; docker run \/bin\/bash -c \\"ls -l\\"/;
       h.expect(outputs[0]).to.match(regex);
     });